chore: disable SDKv2 warning message
The JS SDKv2 is starting to emit a warning message to scare people away:

```
(node:25530) NOTE: We are formalizing our plans to enter AWS SDK for JavaScript (v2) into maintenance mode in 2023.

Please migrate your code to use AWS SDK for JavaScript (v3).
For more information, check the migration guide at https://a.co/7PzMCcy
(Use `node --trace-warnings ...` to show where the warning was created)
```

We are aware, and will migrate eventually. In the mean time, let's not bother our users with this warning.