fix: continue processing stacks after an error has occurred (#278)
Taking a cue from #272, detect an additional error condition and continue regardless of exceptions.

Thanks @mattgresham for the diagnosis and fix of these issues! Rewritten in a slightly different style that is going to catch errors in processing of a single stack as well.

This now finds 22 stacks in my account, where previously it would only find 12.

Also replace the initial JSON parse followed by a YAML parse with a single YAML parse, as YAML is a superset of JSON.

Fixes #273, closes #272.